test: split MockTransport into a separate file
test_bridge is getting pretty big, mostly because a lot of tests want to
use MockTransport.  Put it in its own file.
